restructued zeus, multinest
