+++ autofit/non_linear/mcmc/emcee/plotter.py
@@ -1,4 +1,3 @@
-import matplotlib.pyplot as plt
 import numpy as np
 import corner
