+++ autofit/non_linear/mcmc/zeus/samples.py
@@ -1,29 +1,24 @@
 from typing import Optional
 
-import emcee
 import numpy as np
+import zeus
 
 from autofit import MCMCSamples
 from autofit.mapper.model_mapper import ModelMapper
 from autofit.non_linear.mcmc.auto_correlations import AutoCorrelationsSettings, AutoCorrelations
 from autofit.non_linear.samples import Sample
 
-
-class EmceeSamples(MCMCSamples):
+class ZeusSamples(MCMCSamples):
 
     def __init__(
             self,
             model: ModelMapper,
-            backend: emcee.backends.HDFBackend,
+            zeus_sampler,
             auto_correlation_settings: AutoCorrelationsSettings,
             unconverged_sample_size: int = 100,
             time: Optional[float] = None,
     ):
         """
-        Create a `Samples` object from this non-linear search's output files on the hard-disk and model.
-
-        For Emcee, all quantities are extracted via the hdf5 backend of results.
-
         Attributes
         ----------
         total_walkers : int
@@ -33,21 +28,15 @@ class EmceeSamples(MCMCSamples):
             to the total steps * total walkers).
         """
 
-        self.backend = backend
+        self.zeus_sampler = zeus_sampler
 
-        parameter_lists = self.backend.get_chain(flat=True).tolist()
-
+        parameter_lists = self.zeus_sampler.get_chain(flat=True).tolist()
         log_prior_list = [
             sum(model.log_prior_list_from_vector(vector=vector)) for vector in parameter_lists
         ]
-
-        log_posterior_list = self.backend.get_log_prob(flat=True).tolist()
-
-        log_likelihood_list = [
-            log_posterior - log_prior for
-            log_posterior, log_prior in
-            zip(log_posterior_list, log_prior_list)
-        ]
+        log_posterior_list = self.zeus_sampler.get_log_prob(flat=True).tolist()
+        log_likelihood_list = [log_posterior - log_prior for log_posterior, log_prior in
+                               zip(log_posterior_list, log_prior_list)]
 
         weight_list = len(log_likelihood_list) * [1.0]
 
@@ -69,30 +58,32 @@ class EmceeSamples(MCMCSamples):
 
     @property
     def samples_after_burn_in(self) -> [list]:
-        """
-        The emcee samples with the initial burn-in samples removed.
+        """The zeus samples with the initial burn-in samples removed.
 
-        The burn-in period is estimated using the auto-correlation times of the parameters.
-        """
+        The burn-in period is estimated using the auto-correlation times of the parameters."""
+
         discard = int(3.0 * np.max(self.auto_correlations.times))
         thin = int(np.max(self.auto_correlations.times) / 2.0)
-        return self.backend.get_chain(discard=discard, thin=thin, flat=True)
+        return self.zeus_sampler.get_chain(discard=discard, thin=thin, flat=True)
 
     @property
     def total_walkers(self):
-        return len(self.backend.get_chain()[0, :, 0])
+        return len(self.zeus_sampler.get_chain()[0, :, 0])
 
     @property
     def total_steps(self):
-        return len(self.backend.get_log_prob())
+        return int(self.zeus_sampler.ncall_total)
 
     @property
     def auto_correlations(self):
-        times = self.backend.get_autocorr_time(tol=0)
 
-        previous_auto_correlation_times = emcee.autocorr.integrated_time(
-            x=self.backend.get_chain()[: -self.auto_correlation_settings.check_size, :, :], tol=0
-        )
+        times = zeus.AutoCorrTime(samples=self.zeus_sampler.get_chain())
+        try:
+            previous_auto_correlation_times = zeus.AutoCorrTime(
+                samples=self.zeus_sampler.get_chain()[: -self.auto_correlation_settings.check_size, :, :],
+            )
+        except IndexError:
+            previous_auto_correlation_times = None
 
         return AutoCorrelations(
             check_size=self.auto_correlation_settings.check_size,
