+++ autofit/plot/__init__.py
@@ -1,6 +1,6 @@
 from autofit.plot.samples_plotters import SamplesPlotter
 from autofit.plot.dynesty_plotter import DynestyPlotter
 from autofit.plot.ultranest_plotter import UltraNestPlotter
-from autofit.plot.emcee_plotter import EmceePlotter
-from autofit.plot.zeus_plotter import ZeusPlotter
+from autofit.non_linear.mcmc.emcee.plotter import EmceePlotter
+from autofit.non_linear.mcmc.zeus.plotter import ZeusPlotter
 from autofit.plot.pyswarms_plotter import PySwarmsPlotter
