+++ test_autofit/non_linear/nest/test_multi_nest.py
@@ -8,46 +8,46 @@ import pytest
 import autofit as af
 from autoconf import conf
 from autofit.mock import mock
-from autofit.non_linear.nest import multi_nest as mn
+from autofit.non_linear.nest.multinest import samples as samp
 
 directory = path.dirname(path.realpath(__file__))
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
 
 
-@pytest.fixture(name="multi_nest_summary_path")
-def test_multi_nest_summary():
-    multi_nest_summary_path = path.join(conf.instance.output_path, "non_linear", "multinest", "summary")
+@pytest.fixture(name="multinest_summary_path")
+def test_multinest_summary():
+    multinest_summary_path = path.join(conf.instance.output_path, "non_linear", "multinest", "summary")
 
-    if path.exists(multi_nest_summary_path):
-        shutil.rmtree(multi_nest_summary_path)
+    if path.exists(multinest_summary_path):
+        shutil.rmtree(multinest_summary_path)
 
-    os.mkdir(multi_nest_summary_path)
+    os.mkdir(multinest_summary_path)
 
-    return multi_nest_summary_path
-
-
-@pytest.fixture(name="multi_nest_samples_path")
-def test_multi_nest_samples():
-    multi_nest_samples_path = path.join(conf.instance.output_path, "non_linear", "multinest", "samples")
-
-    if path.exists(multi_nest_samples_path):
-        shutil.rmtree(multi_nest_samples_path)
-
-    os.mkdir(multi_nest_samples_path)
-
-    return multi_nest_samples_path
+    return multinest_summary_path
 
 
-@pytest.fixture(name="multi_nest_resume_path")
-def test_multi_nest_resume():
-    multi_nest_resume_path = path.join(conf.instance.output_path, "non_linear", "multinest", "resume")
+@pytest.fixture(name="multinest_samples_path")
+def test_multinest_samples():
+    multinest_samples_path = path.join(conf.instance.output_path, "non_linear", "multinest", "samples")
 
-    if path.exists(multi_nest_resume_path):
-        shutil.rmtree(multi_nest_resume_path)
+    if path.exists(multinest_samples_path):
+        shutil.rmtree(multinest_samples_path)
 
-    os.mkdir(multi_nest_resume_path)
+    os.mkdir(multinest_samples_path)
 
-    return multi_nest_resume_path
+    return multinest_samples_path
+
+
+@pytest.fixture(name="multinest_resume_path")
+def test_multinest_resume():
+    multinest_resume_path = path.join(conf.instance.output_path, "non_linear", "multinest", "resume")
+
+    if path.exists(multinest_resume_path):
+        shutil.rmtree(multinest_resume_path)
+
+    os.mkdir(multinest_resume_path)
+
+    return multinest_resume_path
 
 
 def create_path(func):
@@ -128,7 +128,7 @@ def create_resume(file_path):
 
 class TestMulitNest:
     def test__loads_from_config_file_if_not_input(self):
-        multi_nest = af.MultiNest(
+        multinest = af.MultiNest(
             prior_passer=af.PriorPasser(sigma=2.0, use_errors=False, use_widths=False),
             n_live_points=40,
             sampling_efficiency=0.5,
@@ -150,40 +150,40 @@ class TestMulitNest:
             init_MPI=True,
         )
 
-        assert multi_nest.prior_passer.sigma == 2.0
-        assert multi_nest.prior_passer.use_errors is False
-        assert multi_nest.prior_passer.use_widths is False
-        assert multi_nest.config_dict_search["n_live_points"] == 40
-        assert multi_nest.config_dict_search["sampling_efficiency"] == 0.5
+        assert multinest.prior_passer.sigma == 2.0
+        assert multinest.prior_passer.use_errors is False
+        assert multinest.prior_passer.use_widths is False
+        assert multinest.config_dict_search["n_live_points"] == 40
+        assert multinest.config_dict_search["sampling_efficiency"] == 0.5
 
-        multi_nest = af.MultiNest()
+        multinest = af.MultiNest()
 
-        assert multi_nest.prior_passer.sigma == 3.0
-        assert multi_nest.prior_passer.use_errors is True
-        assert multi_nest.prior_passer.use_widths is True
-        assert multi_nest.config_dict_search["n_live_points"] == 50
-        assert multi_nest.config_dict_search["sampling_efficiency"] == 0.6
+        assert multinest.prior_passer.sigma == 3.0
+        assert multinest.prior_passer.use_errors is True
+        assert multinest.prior_passer.use_widths is True
+        assert multinest.config_dict_search["n_live_points"] == 50
+        assert multinest.config_dict_search["sampling_efficiency"] == 0.6
 
         model = af.ModelMapper(mock_class_1=mock.MockClassx4)
 
         fitness = af.MultiNest.Fitness(
             analysis=None,
             model=model,
-            samples_from_model=multi_nest.samples_from,
+            samples_from_model=multinest.samples_from,
             stagger_resampling_likelihood=False,
             paths=None
         )
 
         assert fitness.model == model
 
-    def test__read_quantities_from_weighted_samples_file(self, multi_nest_samples_path):
-        multi_nest = af.MultiNest()
-        multi_nest.paths = af.DirectoryPaths(path_prefix=path.join("non_linear", "multinest"))
+    def test__read_quantities_from_weighted_samples_file(self, multinest_samples_path):
+        multinest = af.MultiNest()
+        multinest.paths = af.DirectoryPaths(path_prefix=path.join("non_linear", "multinest"))
 
-        create_weighted_samples_4_parameters(file_path=multi_nest.paths.path)
+        create_weighted_samples_4_parameters(file_path=multinest.paths.path)
 
-        parameters = mn.parameters_from_file_weighted_samples(
-            file_weighted_samples=path.join(multi_nest.paths.path, "multinest.txt"),
+        parameters = samp.parameters_from_file_weighted_samples(
+            file_weighted_samples=path.join(multinest.paths.path, "multinest.txt"),
             prior_count=4,
         )
 
@@ -200,58 +200,58 @@ class TestMulitNest:
             [1.0, 2.0, 3.0, 4.0],
         ]
 
-        log_likelihood_list = mn.log_likelihood_list_from_file_weighted_samples(
-            file_weighted_samples=path.join(multi_nest.paths.path, "multinest.txt")
+        log_likelihood_list = samp.log_likelihood_list_from_file_weighted_samples(
+            file_weighted_samples=path.join(multinest.paths.path, "multinest.txt")
         )
 
         value = -0.5 * 9999999.9
 
         assert log_likelihood_list == 10 * [value]
 
-        weight_list = mn.weight_list_from_file_weighted_samples(
-            file_weighted_samples=path.join(multi_nest.paths.path, "multinest.txt")
+        weight_list = samp.weight_list_from_file_weighted_samples(
+            file_weighted_samples=path.join(multinest.paths.path, "multinest.txt")
         )
 
         assert weight_list == [0.02, 0.02, 0.01, 0.05, 0.1, 0.1, 0.1, 0.1, 0.2, 0.3]
 
-    def test__read_total_samples_from_file_resume(self, multi_nest_resume_path):
-        multi_nest = af.MultiNest()
+    def test__read_total_samples_from_file_resume(self, multinest_resume_path):
+        multinest = af.MultiNest()
 
-        create_resume(file_path=multi_nest.paths.path)
+        create_resume(file_path=multinest.paths.path)
 
-        total_samples = mn.total_samples_from_file_resume(
-            file_resume=path.join(multi_nest.paths.path, "multinestresume.dat")
+        total_samples =samp.total_samples_from_file_resume(
+            file_resume=path.join(multinest.paths.path, "multinestresume.dat")
         )
 
         assert total_samples == 12345
 
-    def test__log_evidence_from_file_summary(self, multi_nest_summary_path):
-        multi_nest = af.MultiNest()
-        multi_nest.paths = af.DirectoryPaths(path_prefix=path.join("non_linear", "multinest"))
+    def test__log_evidence_from_file_summary(self, multinest_summary_path):
+        multinest = af.MultiNest()
+        multinest.paths = af.DirectoryPaths(path_prefix=path.join("non_linear", "multinest"))
 
-        create_summary_4_parameters(file_path=multi_nest.paths.samples_path)
+        create_summary_4_parameters(file_path=multinest.paths.samples_path)
 
-        log_evidence = mn.log_evidence_from_file_summary(
-            file_summary=path.join(multi_nest.paths.samples_path, "multinestsummary.txt"),
+        log_evidence = samp.log_evidence_from_file_summary(
+            file_summary=path.join(multinest.paths.samples_path, "multinestsummary.txt"),
             prior_count=4,
         )
 
         assert log_evidence == 0.02
 
     def test__samples_from_model(
-            self, multi_nest_samples_path, multi_nest_resume_path, multi_nest_summary_path
+            self, multinest_samples_path, multinest_resume_path, multinest_summary_path
     ):
-        multi_nest = af.MultiNest()
-        multi_nest.paths = af.DirectoryPaths(path_prefix=path.join("non_linear", "multinest"))
+        multinest = af.MultiNest()
+        multinest.paths = af.DirectoryPaths(path_prefix=path.join("non_linear", "multinest"))
 
-        create_weighted_samples_4_parameters(file_path=multi_nest.paths.samples_path)
-        create_resume(file_path=multi_nest.paths.samples_path)
-        create_summary_4_parameters(file_path=multi_nest.paths.samples_path)
+        create_weighted_samples_4_parameters(file_path=multinest.paths.samples_path)
+        create_resume(file_path=multinest.paths.samples_path)
+        create_summary_4_parameters(file_path=multinest.paths.samples_path)
 
         model = af.ModelMapper(mock_class=mock.MockClassx4)
         model.mock_class.two = af.LogUniformPrior(lower_limit=1e-8, upper_limit=10.0)
 
-        samples = multi_nest.samples_from(model=model)
+        samples = multinest.samples_from(model=model)
 
         assert samples.parameter_lists == [
             [1.1, 2.1, 3.1, 4.1],
